# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job Seeker API",
    "description": "Dopasowuje oferty pracy z serwisów (na razie JustJoin.it) do CV kandydata.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/sources/{source}/categories": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Categories",
        "operationId": "list_categories_sources__source__categories_get",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryInfo"
                  },
                  "title": "Response List Categories Sources  Source  Categories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sync": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Start Sync",
        "description": "Start downloading offers in the background; poll GET /sync/status for progress.",
        "operationId": "start_sync_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sources"
        ],
        "summary": "Cancel Sync",
        "description": "Stop the running sync; offers downloaded so far are kept.",
        "operationId": "cancel_sync_sync_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
...
    "/syncs": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Sync History",
        "operationId": "sync_history_syncs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SyncHistoryItem"
                  },
                  "title": "Response Sync History Syncs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/preferences": {
      "get": {
        "tags": [
          "sources",
          "meta"
        ],
        "summary": "Preferences",
        "description": "Effective search preferences: config.toml [search], optionally with a named preset applied.",
        "operationId": "preferences_preferences_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchPreferences"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryInfo": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "key",
          "count"
        ],
        "title": "CategoryInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchPreferences": {
        "properties": {
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          },
          "experience_levels": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Seniority"
                },
                "type": "array"
              },
              {
                "type": "string",
                "const": "auto"
              }
            ],
            "title": "Experience Levels",
            "description": "'auto' = your level in the offer's technologies plus one level up; or an explicit list",
            "default": "auto"
          },
          "workplace": {
            "items": {
              "$ref": "#/components/schemas/WorkplaceType"
            },
            "type": "array",
            "title": "Workplace"
          },
          "preferred_cities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Cities"
          },
          "onsite_only_in_preferred_cities": {
            "type": "boolean",
            "title": "Onsite Only In Preferred Cities",
            "description": "Drop hybrid/office offers that are not located in one of preferred_cities",
            "default": true
          },
          "min_salary_pln_month": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Salary Pln Month"
          },
          "exclude_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclude Keywords",
            "description": "Whole words matched in title and skills"
          },
          "title_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Title Keywords",
            "description": "Words that make a job title attractive (core skills from the CV count automatically)"
          },
          "avoid_title_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Avoid Title Keywords",
            "description": "Words that make a job title unattractive (soft penalty, not a filter)"
          },
          "target_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Skills",
            "description": "Skills you are moving towards (e.g. C++): partial credit in matching, told to the AI scorer"
          },
          "max_offer_age_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Offer Age Days",
            "default": 45
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sources",
          "categories",
          "experience_levels",
          "workplace",
          "preferred_cities",
          "onsite_only_in_preferred_cities",
          "min_salary_pln_month",
          "exclude_keywords",
          "title_keywords",
          "avoid_title_keywords",
          "target_skills",
          "max_offer_age_days"
        ],
        "title": "SearchPreferences",
        "description": "What the user is looking for. Every field can be overridden per request (CLI flags / API params)."
      },
...
      "Seniority": {
        "type": "string",
        "enum": [
          "intern",
          "junior",
          "mid",
          "senior",
          "manager",
          "c_level"
        ],
        "title": "Seniority"
      },
...
      "SyncHistoryItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "categories": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categories",
            "description": "Empty: all categories; null: not recorded (older runs)"
          },
          "fetched": {
            "type": "integer",
            "title": "Fetched"
          },
          "new": {
            "type": "integer",
            "title": "New"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "started_at",
          "finished_at",
          "categories",
          "fetched",
          "new",
          "error"
        ],
        "title": "SyncHistoryItem"
      },
      "SyncRequest": {
        "properties": {
          "sources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources",
            "description": "Default: search.sources from config"
          },
          "categories": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categories",
            "description": "Default: search.categories from config"
          },
          "search": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search",
            "description": "Take sources/categories from a named search preset"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          }
        },
        "type": "object",
        "title": "SyncRequest"
      },
      "SyncResultItem": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fetched": {
            "type": "integer",
            "title": "Fetched"
          },
          "new": {
            "type": "integer",
            "title": "New"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "source",
          "fetched",
          "new",
          "error"
        ],
        "title": "SyncResultItem"
      },
      "SyncStatus": {
        "properties": {
          "running": {
            "type": "boolean",
            "title": "Running"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "category_index": {
            "type": "integer",
            "title": "Category Index"
          },
          "category_count": {
            "type": "integer",
            "title": "Category Count"
          },
          "fetched": {
            "type": "integer",
            "title": "Fetched"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "cancelled": {
            "type": "boolean",
            "title": "Cancelled"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SyncResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "running",
          "started_at",
          "finished_at",
          "source",
          "category",
          "category_index",
          "category_count",
          "fetched",
          "total",
          "cancelled",
          "error",
          "results"
        ],
        "title": "SyncStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkplaceType": {
        "type": "string",
        "enum": [
          "remote",
          "hybrid",
          "office"
        ],
        "title": "WorkplaceType"
      }
    }
  }
}